+++ src/server/index.mjs
@@ -1,1 +1,0 @@
-export { default as ssr } from './ssr.js';