+++ src/test/index.js
@@ -4,8 +4,7 @@ const { TestDirector } = require('test-director');
 
 const tests = new TestDirector();
 
-require('./server/ssr.test')(tests);
-require('./universal/arrayFlat.test')(tests);
+require('./arrayFlat.test')(tests);
 require('./universal/GraphQL.test')(tests);
 require('./universal/graphqlFetchOptions.test')(tests);
 require('./universal/hashObject.test')(tests);